fix: keep host auth refresh local
