--- packages/host-service/src/providers/auth/JwtApiAuthProvider/index.ts
@@ -1,6 +1,0 @@
-export {
-	JwtApiAuthProvider,
-	type JwtApiAuthProviderAuthState,
-	type JwtApiAuthProviderExpiredState,
-	type JwtApiAuthProviderOptions,
-} from "./JwtApiAuthProvider";
