+++ packages/host-service/src/providers/auth/JwtAuthProvider/index.ts
@@ -1,5 +1,1 @@
-export {
-	JwtApiAuthProvider,
-	JwtApiAuthProvider as JwtAuthProvider,
-	type JwtApiAuthProviderOptions,
-} from "./JwtAuthProvider";
+export { JwtApiAuthProvider } from "./JwtAuthProvider";
