+++ packages/host-service/src/providers/auth/index.ts
@@ -1,8 +1,4 @@
 export { DeviceKeyApiAuthProvider } from "./DeviceKeyAuthProvider";
 export { SESSION_EXPIRED_HINT } from "./hint";
-export {
-	JwtApiAuthProvider,
-	JwtApiAuthProvider as JwtAuthProvider,
-	type JwtApiAuthProviderOptions,
-} from "./JwtApiAuthProvider";
+export { JwtApiAuthProvider } from "./JwtAuthProvider";
 export type { ApiAuthProvider, AuthSessionEventPublisher } from "./types";
