+++ packages/shared/src/auth/index.ts
@@ -1,3 +1,2 @@
 export * from "./authorization";
 export * from "./roles";
-export * from "./token-refresh";
